requirements.txt
pymupdf==1.26.7
google-generativeai==0.8.6
streamlit==1.42.0
python-dotenv==1.0.1
Pillow==10.4.0
